Script to parse the data into a json file
The json file has the following format
{
"path to image_file" : class_category
}

This is done separately for both training and test sets.
Fold flag determines which fold is used for generating the json file.
Save flag determines if the dictionaries are saved as json files or are
printed on screen.

Signed-off-by: Deepan Das <deepandas11@gmail.com>